restructure for simple structur
