--- src/app/auth/login/page.tsx
@@ -1,9 +1,0 @@
-import Login from "@/components/container/01_Auth/Login";
-
-export default function Home() {
-  return (
-    <>
-    <Login />
-    </>
-  );
-}
