+++ src/app/buyer/after_store/page.tsx
@@ -1,9 +1,9 @@
 "use client";
-import HeroSection from "@/components/container/02_Buyer/Home/Hero";
-import ProductGrid from "@/components/container/02_Buyer/Home/ProductGrid";
-import Footer from "@/components/container/02_Buyer/Home/Footer";
+import HeroSection from "@/components/container/Hero";
+import ProductGrid from "@/components/container/ProductGrid";
+import Footer from "@/components/container/Footer";
 import BuyerAppProvider from "@/lib/providers/buyer_provider";
-import NavbarAfterStore from "@/components/container/02_Buyer/Home/Navbar_after_store";
+import NavbarAfterStore from "@/components/container/Navbar_after_store";
 
 export default function BuyerPage() {
   return (
