+++ src/app/buyer/before_login/page.tsx
@@ -1,9 +1,9 @@
 "use client";
-import HeroSection from "@/components/container/02_Buyer/Home/Hero";
-import ProductGrid from "@/components/container/02_Buyer/Home/ProductGrid";
-import Footer from "@/components/container/02_Buyer/Home/Footer";
+import HeroSection from "@/components/container/Hero";
+import ProductGrid from "@/components/container/ProductGrid";
+import Footer from "@/components/container/Footer";
 import BuyerAppProvider from "@/lib/providers/buyer_provider";
-import NavbarBeforeLogin from "@/components/container/02_Buyer/Home/Navbar_before_login";
+import NavbarBeforeLogin from "@/components/container/Navbar_before_login";
 
 export default function BuyerPage() {
   return (
