+++ src/app/buyer/open_store/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import OpenStore from "@/components/container/02_Buyer/Open_Store/Open_Store";
+import OpenStore from "@/components/container/Open_Store";
 export default function BuyerPage() {
   return (
     <OpenStore />
