+++ src/app/buyer/success_create_store/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import SuccessStore from "@/components/container/02_Buyer/Result_Create_Store/Success_Create_Store";
+import SuccessStore from "@/components/container/Success_Create_Store";
 export default function BuyerPage() {
   return (
     <SuccessStore />
